refactor: rename to CandidateConfirmation
Because we are using responsive design for this component, not use it as adaptive design
Also updated ScheduleInterviewPage
