+++ src/pages/schedule-interview/components/CandidateConfirmation/index.ts
@@ -1,0 +1,2 @@
+export * from "./CandidateConfirmation"
+export * from "./CandidateConfirmationInfo"
