--- src/pages/schedule-interview/components/MobileCandidateConfirmation/index.ts
@@ -1,1 +1,0 @@
-export * from "./MobileCandidateConfirmation"
